refactor: rewire test imports from facade to taxonomy modules
All remaining taxonomy-related tests now import or mock taxonomy-store and taxonomy-selectors directly. category-manager.js has zero remaining importers in the active test suite.
